Iniciando a criação da lista de servicos

+++ lib/src/pages/cadastro_servico/cadastro_servico_state.dart
@@ -72,7 +72,7 @@ class CadastrosServicoForm {
   bool? porcentagemServico;
   String? servicoSelecionando;
   int? porcentagem;
-  int? valor;
+  double? valor;
 
   @JsonKey(name: 'created_at')
   DateTime? createdAt;
